pyproject.toml
[project]
name = "clangen"
version = "0.7.9"
description = "Fan-edit of the warrior cat clangen game"
license = "MPL-2.0"
authors = [
    { name = "Sablesteel", email = "noemail@null.com" },
]
readme = "README.md"
keywords = [
    "game",
    "warrior cat",
    "fan-edit",
    "warrior cats",
]
requires-python = ">=3.10"
dependencies = [
    "black~=23.0",
    "cryptography~=42.0",
    "pgpy~=0.6",
    "platformdirs~=3.3",
    "pydantic>=2.12.0",
    "pygame-ce~=2.2",
    "pygame-gui~=0.6",
    "requests~=2.28",
    "strenum~=0.4",
    "tomli>=2.2.1",
    "ujson~=5.7",
]

[dependency-groups]
build = [
    "pillow~=9.5",
    "pyinstaller~=6.6",
    "setuptools~=67.7",
    "wheel~=0.40",
]
dev = [
    "hatchling>=1.27.0",
    "pytest~=8.2",
]
docs = [
    "mkdocs~=1.6",
    "mkdocs-git-committers-plugin-2>=2.4.1",
    "mkdocs-git-revision-date-localized-plugin>=1.3.0",
    "mkdocs-material~=9.7",
]
test = [
    "pylint>=2.17.3",
]

[project.optional-dependencies]
discord = [
    "pypresence~=4.2",
]

[tool.pytest.ini_options]
pythonpath = [
  "."
]

[project.scripts]
dump_model_schemas = "scripts.dump_model_schemas:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts/"]
